feat: Refine and complete mobile and tablet styles.
Spent quite a bit of time getting the tablet styles correct and putting the elements in the correct structure to make it as easier to style responsively.

Added lenis scroll to be reenabled later.
Added new computer gradient to reflect the lighting from the parent element.
Added wave svg to top and bottom to create cohesiveness.