fix: Reduce Javascript on load.
I deployed and tested the site and it scored a bit poorly on performance, but 100s on everything else. The main culprit was large Javascript size, most likely from R3F Canvas.

The resource below documents a great fix to lazy load the canvas after site is mounted. We don't need it on load because its far down the page, so this could work. Will deploy and test.

Resources:
- Performance Optimization for three.js Web Animations: https://www.gatsbyjs.com/blog/performance-optimization-for-three-js-web-animations/