feat: Add physical glass shader to 3D logos, add perspective to JS section
The transmissive material is applied to some of individual shapes of the logo, I think this adds a lot of depth and makes the 3D animation feel more full and complete.
Previously, it felt quite flat and I also reduced the bloom shader to be an outline. This feels more clean and less like a badly optimized render.
Overall I'm really happy with this feature and how it looks.

Perspective skewing was added to the javascript section to match the mockup and styled for mobile, tablet, desktop.

Will deploy and test perf soon.