refactor/fix: Add scrollY tracking on markup, refactor svg
There was a bug with the SVG for mobile where there was a 1px gap on left and right sides. The solution was to scrap next/image rendering svg and use html svg instead.

ScrollY tracking is achieved by subtracting the section position by the total pixels of the page. This will be used later for parallax similar to the intro. Only next steps left are the scores section and then clean up code.