refactor: Change parallax from scrollY to scrollYProgress on Intro hero.
Previously, we were using the scrollY value to parallax but this was definitive, rather than a relative value.
The problem occurs when we are trying to change the scrollY of layer sections, we have to subtract and set up useEffect to run calculations on the difference each scroll.

This solution uses allows us to use the same code but change the offset on the later sections so we can begin tracking individual sections scroll progress rather than entire scroll progress.